feat: MCP 工具系统 + SSE 流式 + WebSearch + MCP 管理 UI + warning 修复
- MCP 风格工具框架：AiTool(JSON Schema) + ToolRegistry(listToolsJson)
- MCP 远程调用：MCPClient(JSON-RPC 2.0) + MCPBridge
- 4 个内置工具：time/note/recall/send_message（全部 JSON Schema 化）
- WebSearch 工具：DuckDuckGo Instant Answer + HTML Lite 双后端
- SSE 流式：OpenAI + Anthropic 双协议实现
- 拆句改进：\n\n/【顿】双分隔符 + 字数×50ms 动态延迟
- 设置页 MCP 服务管理 UI（添加/删除/测试/刷新）
- 修复 14 条编译 warning，零警告编译
- 增加 Gradle Wrapper（9.5.1），支持 CLI 编译

+++ app/src/main/java/com/aftglw/devapi/network/LocalAiService.kt
@@ -151,7 +151,7 @@ class QwenTokenizer(private val ctx: Context) {
         // 1. UTF-8 encode text → bytes
         // 2. BPE merge → token IDs
         // 简化返回
-        return intArrayOf(1) + text.take(512).map { (it.code % 100000 + 100).toInt() }.toIntArray()
+        return intArrayOf(1) + text.take(512).map { (it.code % 100000 + 100) }.toIntArray()
     }
 
     fun decode(tokenIds: IntArray): String {
